Hopefully, correctly manage world and body memory.
With this scheme a world should only be collected if there
are no external references to it and no external references
to any bodies in it.

The body wrappers are only collected when the body is destroyed
(and there are no more references to the wrapper) or the world
is destroyed (and there are no more references to the wrapper).



git-svn-id: https://lotech.googlecode.com/svn/trunk@33 cc13c597-6222-e281-3f33-90c271915fff